LPS-23769 LastMergeTime is long not integer

+++ portal-impl/src/com/liferay/portal/service/impl/LayoutLocalServiceVirtualLayoutsAdvice.java
@@ -309,7 +309,7 @@ public class LayoutLocalServiceVirtualLayoutsAdvice
 		UnicodeProperties settingsProperties =
 			layoutSet.getSettingsProperties();
 
-		long lastMergeTime = GetterUtil.getInteger(
+		long lastMergeTime = GetterUtil.getLong(
 			settingsProperties.getProperty("lastMergeTime"));
 
 		Date modifiedDate = layoutSetPrototype.getModifiedDate();
